requirements.txt
sponsorblock.py~=0.2.3
feedparser~=6.0.10
moviepy~=1.0.3
yt-dlp~=2024.10.7
python-dotenv~=1.0.1